refactor(rpc): extract parallel scan helpers
Move integer primary-key parallel scan helpers out of lineairdb_rpc.cc into parallel_scan.{cc,hh}.
Register parallel_scan.cc in the server build.
